Move click handler for watchedfilter to the Searchbox view
This patch fixes a bug where the click() behaviour would stop
happening after going to any other route and then back to the
route:movies listing.

This is because the checkbox is rendered only inside route:movies,
which means that the onClick handler setup on $(document).ready()
disapears on a redraw.

The fix is setting the onClick handler whenever we render the
Searchbox view.